requirements/requirements.txt
amqp==2.5.1
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.6.1.0
celery==4.3.0
cffi==1.12.3
cryptography==2.7
Django==2.2.5
django-bootstrap3==11.1.0
django-celery-beat==1.5.0
django-filter==2.2.0
django-ranged-response==0.2.0
django-redis-cache==2.0.0
django-simple-captcha==0.5.12
django-timezone-field==3.0
djangorestframework==3.10.2
djangorestframework-bulk==0.2.1
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.20
kombu==4.6.4
more-itertools==7.2.0
passlib==1.7.1
Pillow==6.1.0
pkg-resources==0.0.0
pycparser==2.19
pyOpenSSL==19.0.0
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2019.2
redis==3.3.8
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
zipp==0.6.0
